refactor(settings): SettingsModal 拆分为容器/分区/编辑器/hook/工具方法
1222 行单文件 SettingsModal 按职责拆开（容器 175 行）：
- hooks/useSettingsDraft —— 草稿 → 整体保存状态机（连接 / LLM / 代理 / 轮询 /
  目录 / 语言），暴露各分区所需的语义化 state 与 setter（编辑即标脏）
- sections/ —— 九个独立分区（语言 / 连接 / 轮询 / LLM / 代理 / Agent 目录 /
  工作目录 / 缓存目录 / 运行环境），各自成文件、可独立维护
- editors/ —— 连接 / 代理 / LLM 编辑器弹窗，统一改用 common/Modal 壳
- elements/UpdateCheckButton —— 「检查更新」小部件
- utils —— formatBytes + 轮询档位（POLLER_TIERS / nearestPollerIdx）

容器只做布局编排 + 装配分区 + 底栏保存。对外接口与界面行为不变。
